+++ src/test/java/projects/local/OpenLocalConstantinNoTables.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package projects;
+package projects.local;
 
 import org.embl.mobie.io.ImageDataFormat;
 import org.embl.mobie.io.SpimDataOpener;
@@ -36,13 +36,13 @@ import org.junit.jupiter.api.Assertions;
 import lombok.extern.slf4j.Slf4j;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.SpimDataException;
+import projects.BaseTest;
 
 @Slf4j
-public class OpenLocalTestData extends BaseTest {
-
+public class OpenLocalConstantinNoTables extends BaseTest {
     @Test
     public void showYX() throws SpimDataException {
-        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/g/emcf/pape/mobie-test-projects", ImageDataFormat.BdvN5);
+        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/Volumes/emcf/pape/jil", ImageDataFormat.BdvN5);
         Assertions.assertEquals(1, getTimePointsSize());
     }
 }
