+++ src/test/java/projects/local/OpenLocalGiuliaMartinCLEM.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package projects;
+package projects.local;
 
 import org.embl.mobie.io.ImageDataFormat;
 import org.embl.mobie.io.SpimDataOpener;
@@ -36,12 +36,14 @@ import org.junit.jupiter.api.Assertions;
 import lombok.extern.slf4j.Slf4j;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.SpimDataException;
+import projects.BaseTest;
 
 @Slf4j
-public class OpenRemoteCentriolesBranch extends BaseTest {
+public class OpenLocalGiuliaMartinCLEM extends BaseTest {
+
     @Test
     public void showYX() throws SpimDataException {
-        this.spimData = (SpimData) new SpimDataOpener().openSpimData("https://github.com/mobie/centrioles-tomo-datasets/tree/grid-test-tomo", ImageDataFormat.BdvN5S3);
+        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/g/emcf/pape/clem-example-project", ImageDataFormat.BdvN5);
         Assertions.assertEquals(1, getTimePointsSize());
     }
 }
