+++ src/test/java/projects/local/OpenLocalJulianNoTables.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package projects;
+package projects.local;
 
 import org.embl.mobie.io.ImageDataFormat;
 import org.embl.mobie.io.SpimDataOpener;
@@ -36,12 +36,14 @@ import org.junit.jupiter.api.Assertions;
 import lombok.extern.slf4j.Slf4j;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.SpimDataException;
+import projects.BaseTest;
 
 @Slf4j
-public class OpenRemoteZebrafish extends BaseTest {
+public class OpenLocalJulianNoTables extends BaseTest {
+
     @Test
     public void showYX() throws SpimDataException {
-        this.spimData = (SpimData) new SpimDataOpener().openSpimData("https://github.com/mobie/zebrafish-lm-datasets/tree/main", ImageDataFormat.BdvN5S3);
+        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/Volumes/emcf/hennies/for_constantin/mobie_no_table_test/", ImageDataFormat.BdvN5);
         Assertions.assertEquals(1, getTimePointsSize());
     }
 }
