+++ src/test/java/projects/local/OpenLocalTest.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package projects;
+package projects.local;
 
 import org.embl.mobie.io.ImageDataFormat;
 import org.embl.mobie.io.SpimDataOpener;
@@ -36,12 +36,14 @@ import org.junit.jupiter.api.Assertions;
 import lombok.extern.slf4j.Slf4j;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.SpimDataException;
+import projects.BaseTest;
 
 @Slf4j
-public class OpenLocalCovidEMCF extends BaseTest {
+public class OpenLocalTest extends BaseTest {
     @Test
     public void showYX() throws SpimDataException {
-        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/Volumes/emcf/common/5792_Sars-Cov-2/covid-em/data", ImageDataFormat.BdvN5);
+        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/Users/tischer/Desktop/mobie/MartinFloat", ImageDataFormat.OmeZarr);
         Assertions.assertEquals(1, getTimePointsSize());
     }
+
 }
