+++ src/test/java/projects/local/OpenLocalTestData.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package projects;
+package projects.local;
 
 import org.embl.mobie.io.ImageDataFormat;
 import org.embl.mobie.io.SpimDataOpener;
@@ -36,15 +36,14 @@ import org.junit.jupiter.api.Assertions;
 import lombok.extern.slf4j.Slf4j;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.SpimDataException;
+import projects.BaseTest;
 
 @Slf4j
-public class OpenLocalPlankton extends BaseTest {
+public class OpenLocalTestData extends BaseTest {
 
     @Test
     public void showYX() throws SpimDataException {
-        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/Volumes/emcf/pape/plankton-fibsem-project", ImageDataFormat.BdvN5);
+        this.spimData = (SpimData) new SpimDataOpener().openSpimData("/g/emcf/pape/mobie-test-projects", ImageDataFormat.BdvN5);
         Assertions.assertEquals(1, getTimePointsSize());
     }
-//dataset("galdieria"));
-
 }
