+++ src/test/java/projects/remote/OpenRemoteOpenOrganelle.java
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package projects;
+package projects.remote;
 
 import org.embl.mobie.io.ImageDataFormat;
 import org.embl.mobie.io.SpimDataOpener;
@@ -36,12 +36,13 @@ import org.junit.jupiter.api.Assertions;
 import lombok.extern.slf4j.Slf4j;
 import mpicbg.spim.data.SpimData;
 import mpicbg.spim.data.SpimDataException;
+import projects.BaseTest;
 
 @Slf4j
-public class OpenRemoteCovidTomogramsBranch extends BaseTest {
+public class OpenRemoteOpenOrganelle extends BaseTest {
     @Test
     public void showYX() throws SpimDataException {
-        this.spimData = (SpimData) new SpimDataOpener().openSpimData("https://github.com/mobie/covid-tomo-datasets/tree/new-spec", ImageDataFormat.BdvN5S3);
+        this.spimData = (SpimData) new SpimDataOpener().openSpimData("https://github.com/mobie/open-organelle-test", ImageDataFormat.OpenOrganelleS3);
         Assertions.assertEquals(1, getTimePointsSize());
     }
 }
